feat: Add exercise 76
Move partitions functions into separate file, since we're using them
in (at least) exercise 76 and 78.
